[autoscaler] Refactoring to address code review comments.
