+++ python/ray/tests/aws/test_autoscaler_aws.py
@@ -1,0 +1,74 @@
+import pytest
+import copy
+
+import ray.autoscaler.aws.config
+import ray.tests.aws.utils.stubs as stubs
+from ray.tests.aws.utils.constants import AUX_SUBNET, DEFAULT_SUBNET, \
+    DEFAULT_SG_AUX_SUBNET, CONFIG_DEFAULT_RULES_DIFFERENT_VPC, DEFAULT_SG, \
+    DEFAULT_SG_DUAL_GROUP_RULES, DEFAULT_SG_WITH_RULES_AUX_SUBNET, \
+    DEFAULT_SG_WITH_RULES, AUX_SG
+
+
+def test_create_sg_different_vpc_same_rules(iam_client_stub, ec2_client_stub):
+    # use default stubs to skip ahead to security group configuration
+    stubs.skip_to_configure_sg(ec2_client_stub, iam_client_stub)
+
+    # given head and worker nodes with custom subnets defined...
+    # expect to first describe the worker subnet ID
+    stubs.describe_subnets_echo(ec2_client_stub, AUX_SUBNET)
+    # expect to second describe the head subnet ID
+    stubs.describe_subnets_echo(ec2_client_stub, DEFAULT_SUBNET)
+    # given no existing security groups within the VPC...
+    stubs.describe_no_security_groups(ec2_client_stub)
+    # expect to first create a security group on the worker node VPC
+    stubs.create_sg_echo(ec2_client_stub, DEFAULT_SG_AUX_SUBNET)
+    # expect new worker security group details to be retrieved after creation
+    stubs.describe_sgs_on_vpc(ec2_client_stub, [AUX_SUBNET["VpcId"]],
+                              [DEFAULT_SG_AUX_SUBNET])
+    # expect to tag the worker security group after creating it
+    stubs.create_sg_tags(ec2_client_stub, DEFAULT_SG_AUX_SUBNET["GroupId"],
+                         CONFIG_DEFAULT_RULES_DIFFERENT_VPC["cluster_name"])
+    # expect to second create a security group on the head node VPC
+    stubs.create_sg_echo(ec2_client_stub, DEFAULT_SG)
+    # expect new head security group details to be retrieved after creation
+    stubs.describe_sgs_on_vpc(ec2_client_stub, [DEFAULT_SUBNET["VpcId"]],
+                              [DEFAULT_SG])
+    # expect to tag the head security group after creating it
+    stubs.create_sg_tags(ec2_client_stub, DEFAULT_SG["GroupId"],
+                         CONFIG_DEFAULT_RULES_DIFFERENT_VPC["cluster_name"])
+
+    # given no existing default head security group inbound rules...
+    # expect to authorize all default head inbound rules
+    stubs.authorize_sg_ingress(ec2_client_stub, DEFAULT_SG_DUAL_GROUP_RULES)
+    # given no existing default worker security group inbound rules...
+    # expect to authorize all default worker inbound rules
+    stubs.authorize_sg_ingress(ec2_client_stub,
+                               DEFAULT_SG_WITH_RULES_AUX_SUBNET)
+
+    # given the prior modification to the head security group...
+    # expect the next read of a head security group property to reload it
+    stubs.describe_sg_echo(ec2_client_stub, DEFAULT_SG_WITH_RULES)
+    # given the prior modification to the worker security group...
+    # expect the next read of a worker security group property to reload it
+    stubs.describe_sg_echo(ec2_client_stub, DEFAULT_SG_WITH_RULES_AUX_SUBNET)
+
+    # given our mocks and test config as input...
+    # expect the test config to be bootstrapped successfully
+    config = ray.autoscaler.aws.config.bootstrap_aws(
+        copy.deepcopy(CONFIG_DEFAULT_RULES_DIFFERENT_VPC))
+
+    # expect the bootstrapped config to show different head and worker security
+    # groups residing on different subnets
+    assert config["head_node"]["SecurityGroupIds"] == [DEFAULT_SG["GroupId"]]
+    assert config["head_node"]["SubnetIds"] == [DEFAULT_SUBNET["SubnetId"]]
+    assert config["worker_nodes"]["SecurityGroupIds"] == [AUX_SG["GroupId"]]
+    assert config["worker_nodes"]["SubnetIds"] == [AUX_SUBNET["SubnetId"]]
+
+    # expect no pending responses left in IAM or EC2 client stub queues
+    iam_client_stub.assert_no_pending_responses()
+    ec2_client_stub.assert_no_pending_responses()
+
+
+if __name__ == "__main__":
+    import sys
+    sys.exit(pytest.main(["-v", __file__]))
